Only keep cursor out of regular checkboxes
Keeping the cursor out of regular checkboxes (`[ ]` and `[x]`) does not
allow the user to edit the contents directly, but this is not an issue
because one can use the "Toggle checkbox status" command (bound to Cmd-L
or Ctrl-L by default).

On the other hand, keeping the cursor out of any custom checkboxes (e.g.
`[!]`) prevents the user from editing it and causes issues.

Moreover, we were previously considering e.g. `[[]` a custom checkbox
(with character `[` inside the "box"), which caused additional problems
when a link is added just after a bullet point (`[[]]`).